C-Shell Terminal
In this code, terminal structure has written and it needs still some implementations. For now, command split and taking max. 3 param in a command line has been decided and done like that. History is written by creating "history_of_terminal.txt" file and writing in it. If user write "exit" in  a command line, then execution will be done. If command name is not true, then user will see a text which is "That command line is not allowed.".